添加语速调节

+++ app/src/main/java/com/monke/monkeybook/help/ReadBookControl.java
@@ -329,7 +329,7 @@ public class ReadBookControl {
         editor.apply();
     }
 
-    public float getSpeechRate() {
+    public int getSpeechRate() {
         return speechRate;
     }
